product screen + product details screen + cart screen
The user can click on the product section in the drawer to navigate to the product screen. Users can view all of the products on the product screen. Users can choose the category of the product at the horizontal scroll bar to show only the product with the specified product type. Users can search for the product from the search bar. User can view the product details by clicking the product card and the user will be able to view the product details at the product details screen. Users can add the product item to the cart from the product screen and also from the product details screen. Users can specify the quantity of product that they want to add to cart. Users can click on the cart icon to view the product items that they have added to cart. Users can update the quantity of the product items and also remove the product items from their cart.